refactor: Rename src/config/ to src/Config/
