+++ src/File.ts
@@ -1,7 +1,7 @@
 import { MetadataCache, TFile, Vault } from 'obsidian';
 import type { ListItemCache } from 'obsidian';
 
-import { getSettings } from './config/Settings';
+import { getSettings } from './Config/Settings';
 import type { Task } from './Task';
 
 let metadataCache: MetadataCache | undefined;
