+++ src/Query/Sort.ts
@@ -1,6 +1,6 @@
 import type moment from 'moment';
 import type { Task } from '../Task';
-import { getSettings } from '../config/Settings';
+import { getSettings } from '../Config/Settings';
 import type { Query, SortingProperty } from './Query';
 
 type Comparator = (a: Task, b: Task) => number;
